Update ejercicio13.pde
upsis, colgué que tenia mas sentido en float

+++ ejercicio13/ejercicio13.pde
@@ -1,7 +1,7 @@
-int base = 23;
-int altura = 15;
-int perimetro = 2*(base+ altura);
-int area = base * altura;
+float base = 23;
+float altura = 15;
+float perimetro = 2*(base+ altura);
+float area = base * altura;
 
 println("el perimetro del rectangulo es: " + perimetro);
 println("el area del rectangulo es: " + area);
